refactor: seperate openapi-ts client from parser client
