+++ packages/openapi-ts/src/generate/__tests__/services.spec.ts
@@ -4,7 +4,7 @@ import path from 'node:path';
 import { describe, expect, it, vi } from 'vitest';
 
 import { TypeScriptFile } from '../../compiler';
-import type { Operation } from '../../openApi';
+import type { Operation } from '../../types/client';
 import type { Files } from '../../types/utils';
 import { setConfig } from '../../utils/config';
 import { generateServices } from '../services';
