+++ packages/openapi-ts/src/openApi/v2/index.ts
@@ -1,25 +1,25 @@
-import type { Client } from '../../types/client';
+import type { Client } from '../common/interfaces/client';
 import { getServiceVersion } from '../common/parser/service';
 import type { OpenApi } from './interfaces/OpenApi';
 import { getModels } from './parser/getModels';
+import { getOperations } from './parser/getOperations';
 import { getServer } from './parser/getServer';
-import { getServices } from './parser/getServices';
 
 /**
  * Parse the OpenAPI specification to a Client model that contains
- * all the models, services and schema's we should output.
+ * all the models, operations and schema's we should output.
  * @param openApi The OpenAPI spec that we have loaded from disk.
  */
 export const parse = (openApi: OpenApi): Client => {
   const version = getServiceVersion(openApi.info.version);
   const server = getServer(openApi);
   const { models, types } = getModels(openApi);
-  const services = getServices({ openApi, types });
+  const operations = getOperations({ openApi, types });
 
   return {
     models,
+    operations,
     server,
-    services,
     types,
     version,
   };
