+++ packages/openapi-ts/src/openApi/v3/parser/operation.ts
@@ -6,7 +6,6 @@ import type {
 } from '../../common/interfaces/client';
 import { getRef } from '../../common/parser/getRef';
 import { getOperationResponseHeader } from '../../common/parser/operation';
-import { getServiceName } from '../../common/parser/service';
 import { toSortedByRequired } from '../../common/parser/sort';
 import { getConfig } from '../../config';
 import type { OpenApi } from '../interfaces/OpenApi';
@@ -43,7 +42,6 @@ export const getOperation = ({
   op,
   openApi,
   pathParams,
-  tag,
   types,
   url,
 }: {
@@ -52,12 +50,10 @@ export const getOperation = ({
   op: OpenApiOperation;
   openApi: OpenApi;
   pathParams: OperationParameters;
-  tag: string;
   types: Client['types'];
   url: string;
 }): Operation => {
   const config = getConfig();
-  const service = getServiceName(tag);
 
   const operationWithoutName: Omit<Operation, 'name'> = {
     $refs: [],
@@ -76,8 +72,8 @@ export const getOperation = ({
     path: url,
     responseHeader: null,
     responses: [],
-    service,
     summary: op.summary || null,
+    tags: op.tags || null,
   };
   const operation = {
     ...operationWithoutName,
